requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
certifi==2025.8.3
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
dask==2025.7.0
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
executing==2.2.1
fonttools==4.59.2
fsspec==2025.7.0
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imbalanced-learn==0.14.0
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib>=1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lightgbm==4.6.0
locket==1.0.0
matplotlib>=3.10.6
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
parso==0.8.5
partd==1.4.2
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn>=0.13.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
threadpoolctl==3.6.0
toolz==1.0.0
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.13
xgboost==3.0.4
catboost>=1.2
kafka-python
zipp==3.23.0
shap
groq
redis